Add v4vRefreshConnectedProviders to ensure boostagram controls appear on PlayerScreen
#1701

Previously I only was initializing the v4v providers on app launch, but if the network request to the v4v provider failed for any reason, then the app would run as if v4v was not even enabled, and would not render the boost or streaming buttons until the app is restarted. To help ensure the UI renders accurately, I am calling v4vRefreshConnectedProviders in the PlayerScreen componentDidMount, to ensure the buttons render properly if a valid v4v provider connection can be established.