Workaround for RNTP setRepeatMode bug when a single track is loaded
I started trying to add a workaround for buggy RepeatMode handling when only a single track is loaded, but it seems like this might be fixed when we upgrade above RNTP 4.0.0-rc.4, so I figured I'd commit my changes but probably not use them. Wait and see if the upgrade fixes it.

https://github.com/doublesymmetry/react-native-track-player/issues/2070